Move login and signup to pages folder
